tests: move config tests to their own module
